refactor: introduce OpaqueBlockBridge; use it for Raven instead of SSMBlockBridge
