+++ transformer_lens/model_bridge/generalized_components/ssm_block.py
@@ -1,98 +1,26 @@
 """Block container for State Space Model (Mamba) layers: norm → mixer → residual."""
 from __future__ import annotations
 
-import re
-from typing import Any, Dict, Optional
-
-import torch
-
-from transformer_lens.model_bridge.exceptions import StopAtLayerException
-from transformer_lens.model_bridge.generalized_components.base import (
-    GeneralizedComponent,
+from transformer_lens.model_bridge.generalized_components.opaque_block import (
+    OpaqueBlockBridge,
 )
 
 
-class SSMBlockBridge(GeneralizedComponent):
-    """Block bridge for SSM layers — direct GeneralizedComponent subclass.
+class SSMBlockBridge(OpaqueBlockBridge):
+    """Block bridge for SSM layers (Mamba, Falcon-H1, etc.).
 
-    Does not inherit from BlockBridge because BlockBridge's hook_aliases hardcode
-    transformer-specific names (hook_attn_*, hook_mlp_*, hook_resid_mid).
+    Extends :class:`OpaqueBlockBridge` with SSM-specific hook aliases:
+
+    - ``hook_mixer_in``  → ``mixer.hook_in``
+    - ``hook_mixer_out`` → ``mixer.hook_out``
+
+    These aliases are only meaningful for architectures whose blocks contain a
+    ``mixer`` submodule (SSM or hybrid SSM/attention). For non-SSM architectures
+    use :class:`OpaqueBlockBridge` directly.
     """
 
-    is_list_item: bool = True
     hook_aliases = {
-        "hook_resid_pre": "hook_in",
-        "hook_resid_post": "hook_out",
+        **OpaqueBlockBridge.hook_aliases,
         "hook_mixer_in": "mixer.hook_in",
         "hook_mixer_out": "mixer.hook_out",
     }
-
-    def __init__(
-        self,
-        name: str,
-        config: Optional[Any] = None,
-        submodules: Optional[Dict[str, GeneralizedComponent]] = None,
-        hook_alias_overrides: Optional[Dict[str, str]] = None,
-    ):
-        super().__init__(
-            name,
-            config,
-            submodules=submodules if submodules is not None else {},
-            hook_alias_overrides=hook_alias_overrides,
-        )
-
-    def forward(self, *args: Any, **kwargs: Any) -> Any:
-        """Delegate to the HF block with hook_in/hook_out wrapped around it."""
-        if self.original_component is None:
-            raise RuntimeError(
-                f"Original component not set for {self.name}. "
-                "Call set_original_component() first."
-            )
-
-        self._check_stop_at_layer(*args, **kwargs)
-        args, kwargs = self._hook_input_hidden_states(args, kwargs)
-        output = self.original_component(*args, **kwargs)
-        return self._apply_output_hook(output)
-
-    def _apply_output_hook(self, output: Any) -> Any:
-        """Hook the primary output tensor, preserving tuple structure if present."""
-        if isinstance(output, tuple) and len(output) > 0:
-            first = output[0]
-            if isinstance(first, torch.Tensor):
-                first = self.hook_out(first)
-                return (first,) + output[1:]
-            return output
-        if isinstance(output, torch.Tensor):
-            return self.hook_out(output)
-        return output
-
-    def _hook_input_hidden_states(self, args: tuple, kwargs: dict) -> tuple[tuple, dict]:
-        """Hook the hidden_states input whether it arrives positionally or by name."""
-        if len(args) > 0 and isinstance(args[0], torch.Tensor):
-            hooked = self.hook_in(args[0])
-            args = (hooked,) + args[1:]
-        elif "hidden_states" in kwargs and isinstance(kwargs["hidden_states"], torch.Tensor):
-            kwargs["hidden_states"] = self.hook_in(kwargs["hidden_states"])
-        return args, kwargs
-
-    def _check_stop_at_layer(self, *args: Any, **kwargs: Any) -> None:
-        """Raise StopAtLayerException when the configured stop index matches this block."""
-        if not (hasattr(self, "_stop_at_layer_idx") and self._stop_at_layer_idx is not None):
-            return
-        if self.name is None:
-            return
-        # Mamba uses `.layers.{i}`; `blocks.{i}` is the fallback TL convention.
-        match = re.search(r"\.layers\.(\d+)", self.name) or re.search(r"blocks\.(\d+)", self.name)
-        if not match:
-            return
-        layer_idx = int(match.group(1))
-        if layer_idx != self._stop_at_layer_idx:
-            return
-        if len(args) > 0 and isinstance(args[0], torch.Tensor):
-            input_tensor = args[0]
-        elif "hidden_states" in kwargs and isinstance(kwargs["hidden_states"], torch.Tensor):
-            input_tensor = kwargs["hidden_states"]
-        else:
-            raise ValueError(f"Cannot find input tensor to stop at layer {layer_idx}")
-        input_tensor = self.hook_in(input_tensor)
-        raise StopAtLayerException(input_tensor)
